+++ index.html
@@ -3,10 +3,10 @@
 <head>
     <meta charset="UTF-8">
     <title>Title</title>
+    <script defer src="dist/main.js"></script>
 </head>
 <body>
 <link rel='stylesheet' href='dist/style.css'>
-<script src="dist/main.js"></script>
 <div class="searchNew">
 <div class="search-sid">
     <h2>Поиск товаров по артикулам</h2>
